Added comment in the score update / add method - trying to explain approach code has taken.
One refactor could be to replace code sections that are the same in to their own methods and just call them as needed in the processing tree.

The combinations of incoming frame types and then the break down by previous number of games would not change. But by putting the code in to seperate methods for each identical circumstance - the code in the update / add score method would be significantly reduced.